Rename inviteUsers.js -> invitations.js
